Move struct 'AppchainDelegator' and 'AppchainValidator' to separated source files.
