+++ octopus-relay/src/appchain/mod.rs
@@ -1,2 +1,4 @@
+mod delegator;
+mod validator;
 pub mod metadata;
 pub mod state;
