+++ octopus-relay/src/appchain/state.rs
@@ -3,43 +3,13 @@ use std::convert::TryInto;
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::collections::{LazyOption, UnorderedMap, Vector};
 use near_sdk::json_types::U128;
-use near_sdk::{env, AccountId, Balance, BlockHeight, Timestamp};
+use near_sdk::{env, AccountId, Balance, Timestamp};
 
 use crate::storage_key::StorageKey;
-use crate::types::{
-    AppchainStatus, Delegator, Fact, LiteValidator, Locked, Validator, ValidatorSet,
-};
-use crate::{AppchainId, DelegatorId, SeqNum, ValidatorId};
+use crate::types::{AppchainStatus, Fact, LiteValidator, Locked, ValidatorSet};
+use crate::{AppchainId, SeqNum, ValidatorId};
 
-const INVALID_DELEGATORS_DATA_OF_VALIDATOR: &'static str = "Invalid delegators data of validator";
-
-/// Appchain delegator of an appchain validator
-#[derive(BorshDeserialize, BorshSerialize)]
-pub struct AppchainDelegator {
-    /// Id of appchain delegator
-    pub delegator_id: DelegatorId,
-    /// Account id of the delegator
-    pub account_id: AccountId,
-    /// Authorized balance of the delegator
-    pub amount: Balance,
-    /// Block height of the delegation of the delegator
-    pub block_height: BlockHeight,
-}
-
-/// Appchain validator of an appchain
-#[derive(BorshDeserialize, BorshSerialize)]
-pub struct AppchainValidator {
-    /// Id of appchain validator
-    pub validator_id: ValidatorId,
-    /// Account id of the validator
-    pub account_id: AccountId,
-    /// Staked balance of the validator
-    pub amount: Balance,
-    /// Block height which the validator started staking
-    pub block_height: BlockHeight,
-    /// Delegators of the validator
-    pub delegators: UnorderedMap<DelegatorId, LazyOption<AppchainDelegator>>,
-}
+use super::validator::AppchainValidator;
 
 /// Appchain state of an appchain of Octopus Network
 #[derive(BorshDeserialize, BorshSerialize)]
@@ -78,86 +48,6 @@ pub struct AppchainState {
     pub upvote_balance: Balance,
     /// Total downvote balance of OCT token of the appchain
     pub downvote_balance: Balance,
-}
-
-impl AppchainDelegator {
-    /// Convert to struct `Delegator`
-    pub fn to_delegator(&self) -> Delegator {
-        Delegator {
-            id: self.delegator_id.clone(),
-            account_id: self.account_id.clone(),
-            amount: self.amount.into(),
-            block_height: self.block_height,
-        }
-    }
-}
-
-impl AppchainValidator {
-    /// Convert to struct `Validator`
-    pub fn to_validator(&self) -> Validator {
-        Validator {
-            id: self.validator_id.clone(),
-            account_id: self.account_id.clone(),
-            staked_amount: self.amount.into(),
-            block_height: self.block_height,
-            delegators: self
-                .delegators
-                .values_as_vector()
-                .iter()
-                .map(|d| {
-                    d.get()
-                        .expect(INVALID_DELEGATORS_DATA_OF_VALIDATOR)
-                        .to_delegator()
-                })
-                .collect(),
-        }
-    }
-    /// Convert to struct `LiteValidator`
-    pub fn to_lite_validator(&self) -> LiteValidator {
-        LiteValidator {
-            id: self.validator_id.clone(),
-            account_id: self.account_id.clone(),
-            weight: self.amount.into(),
-            block_height: self.block_height,
-            delegators: self
-                .delegators
-                .values_as_vector()
-                .iter()
-                .map(|d| {
-                    d.get()
-                        .expect(INVALID_DELEGATORS_DATA_OF_VALIDATOR)
-                        .to_delegator()
-                })
-                .collect(),
-        }
-    }
-    /// Get delegator by `DelegatorId`
-    pub fn get_delegator(&self, delegator_id: &DelegatorId) -> Option<AppchainDelegator> {
-        if let Some(appchain_delegator_option) = self.delegators.get(delegator_id) {
-            return appchain_delegator_option.get();
-        }
-        Option::None
-    }
-    /// Get total staked amount of OCT tokens of the validator,
-    /// this function will also count all balances of delegators.
-    pub fn get_staked_balance_including_delegators(&self) -> Balance {
-        self.amount
-            + self
-                .delegators
-                .values_as_vector()
-                .iter()
-                .filter(|d| d.is_some())
-                .map(|d| d.get().unwrap().amount)
-                .sum::<u128>()
-    }
-    /// Clear extra storage used by the validator
-    ///
-    /// **This function must be called before remove `AppchainValidator` from storage**
-    pub fn clear_extra_storage(&self) {
-        self.delegators.values_as_vector().iter().for_each(|mut d| {
-            d.remove();
-        });
-    }
 }
 
 impl AppchainState {
@@ -268,7 +158,8 @@ impl AppchainState {
         self.status = AppchainStatus::Booting;
         self.validators_timestamp = env::block_timestamp();
         self.booting_timestamp = env::block_timestamp();
-        self.facts.push(&Fact::UpdateValidatorSet(self.get_latest_validator_set()));
+        self.facts
+            .push(&Fact::UpdateValidatorSet(self.get_latest_validator_set()));
     }
     /// Stake some OCT tokens to the appchain
     pub fn stake(&mut self, validator_id: &ValidatorId, amount: &Balance) -> bool {
